+++ x64barebones/Kernel/include/semaphore.h
@@ -18,7 +18,7 @@ typedef struct semaphore{
 
 
 void semaphore_init(void* ptr);
-semaphore* my_sem_create(uint8_t id, uint64_t value);
+semaphore* my_sem_create(uint64_t id, uint64_t value);
 void my_sem_destroy(semaphore * ptr);
 semaphore* my_sem_open(uint8_t id);
 uint8_t my_sem_post(semaphore *ptr);
